Move modules into utils, io, and testing folders
